add details option for each item

+++ item/models.py
@@ -3,17 +3,19 @@ from django.db import models
 
 class Category(models.Model):
     name = models.CharField(max_length=255)
+
     class Meta:
         ordering = ('name',)
         verbose_name_plural = 'Categories'
     
     def __str__(self):
         return self.name
-class Item (models.Model):
+
+class Item(models.Model):
     category = models.ForeignKey(Category, related_name='items', on_delete=models.CASCADE)
     name = models.CharField(max_length=255)
     description = models.TextField(blank=True, null=True)
-    price = models.DecimalField(max_digits=6, decimal_places=2)
+    price = models.FloatField()
     image = models.ImageField(upload_to='item_images', blank=True, null=True)
     is_sold = models.BooleanField(default=False)
     created_by = models.ForeignKey(User, related_name='items', on_delete=models.CASCADE)
